Revert "refactor(test): impl utils for Droplets/NonDroplets"
This reverts commit 44499f8d0143a128fc6199a497ab8a09104073e7.
