+++ sn_node/tests/msgs_over_gossipsub.rs
@@ -8,6 +8,7 @@
 
 mod common;
 
+use common::NODE_COUNT;
 use eyre::Result;
 use sn_logging::LogBuilder;
 use sn_node::NodeEvent;
